Added system to obtain basic schema information for sqlite
Added a set of interface and implementation to obtain various schema related
information from sqlite. Find out the number of tables, indexes, triggers,
and views. As well as information about the columns and relationships between
them. The system only returns the main schema at current, though later can
enhance to get all the schema's in the system. Currently the design of the
interfaces are very sqlite specific. As more databases are added it can
become more general