[bug] Upgrade will now report if it does not find a version
Before if the initial version the upgrade was not in the database, the upgrade function assumed that the database
was fully upgraded. This is incorrect, in reality this
indicates that there is an issue with the database, or the
migrations being used. Maybe a migration database mismatch.

So, now we will issue an error; if the initial database version
for the upgrade is not found.