[feature] Ability to exclude a table
Now if you only have table that are excluded or the special
table '*' then all the tables except for the excluded table
will be generated. Allow one to build an exclusion list
instead of a inclusion list.

Fixed issue with hidden value of not being just 0 or 1. When the column
is a generated column hidden value will be 2, or 3 (virtual,stored).