[Feature] No need for sequence file if dir only has one sql file
If a directory only has one and only one sql or sql template then a
sequence.txt file is no longer needed. And the sql or sql template file will be applied as if
there was a sequence.txt file with only that sql or sql template file in it.